feat(SPV-1398): move keys package to v2
